[Project] Make variant groups for localized files in add_file_reference
When adding a file reference for a file in a *.lproj folder, use a
variant group to group it with corresponding files for other languages.

Grouping is done by the names without extensions of the files inside the
*.lproj folders.
For example: Base.lproj/Foo.storyboard and de.lproj/Foo.strings will
be put together in a variant group called Foo.